+++ src/BootstrapTable.js
@@ -97,6 +97,7 @@ class BootstrapTable extends Component {
       keyField: keyField,
       colInfos: this.colInfos,
       multiColumnSearch: props.multiColumnSearch,
+      strictSearch: props.strictSearch,
       multiColumnSort: props.multiColumnSort,
       remote: this.props.remote
     });
@@ -177,10 +178,11 @@ class BootstrapTable extends Component {
   }
 
   reset() {
+    const { pageStartIndex } = this.props.options;
     this.store.clean();
     this.setState({
       data: this.getTableData(),
-      currPage: 1,
+      currPage: Util.getFirstPage(pageStartIndex),
       expanding: [],
       sizePerPage: Const.SIZE_PER_PAGE_LIST[0],
       selectedRowKeys: this.store.getSelectedRowKeys(),
@@ -237,6 +239,14 @@ class BootstrapTable extends Component {
         currPage: page,
         sizePerPage,
         reset: false
+      });
+    }
+
+    // If setting the expanded rows is being handled externally
+    // then overwrite the current expanded rows.
+    if (this.props.options.expanding !== options.expanding) {
+      this.setState({
+        expanding: options.expanding || []
       });
     }
 
@@ -403,6 +413,7 @@ class BootstrapTable extends Component {
             withoutNoDataText={ this.props.options.withoutNoDataText }
             expanding={ this.state.expanding }
             onExpand={ this.handleExpandRow }
+            onlyOneExpanding={ this.props.options.onlyOneExpanding }
             beforeShowError={ this.props.options.beforeShowError }
             keyBoardNav={ this.props.keyBoardNav }
             onNavigateCell={ this.handleNavigateCell }
@@ -491,18 +502,7 @@ class BootstrapTable extends Component {
       return;
     }
 
-    // We calculate an offset here in order to properly fetch the indexed data,
-    // despite the page start index not always being 1
-    let normalizedPage;
-    if (pageStartIndex !== undefined) {
-      const offset = Math.abs(Const.PAGE_START_INDEX - pageStartIndex);
-      normalizedPage = page + offset;
-    } else {
-      normalizedPage = page;
-    }
-
-    const result = this.store.page(normalizedPage, sizePerPage).get();
-
+    const result = this.store.page(Util.getNormalizedPage(pageStartIndex, page), sizePerPage).get();
     this.setState({ data: result, reset: false });
   }
 
@@ -535,8 +535,6 @@ class BootstrapTable extends Component {
     let { x, y, currPage } = this.state;
     x += offSetX;
     y += offSetY;
-    // currPage += 1;
-    // console.log(currPage);
 
     const columns = this.store.getColInfos();
     const visibleRowSize = this.state.data.length;
@@ -659,16 +657,17 @@ class BootstrapTable extends Component {
 
   handleShowOnlySelected = () => {
     this.store.ignoreNonSelected();
+    const { pageStartIndex } = this.props.options;
     let result;
     if (this.props.pagination) {
-      result = this.store.page(1, this.state.sizePerPage).get();
+      result = this.store.page(Util.getNormalizedPage(pageStartIndex), this.state.sizePerPage).get();
     } else {
       result = this.store.get();
     }
     this.setState({
       data: result,
       reset: false,
-      currPage: this.props.options.pageStartIndex || Const.PAGE_START_INDEX
+      currPage: Util.getFirstPage(pageStartIndex)
     });
   }
 
@@ -851,14 +850,14 @@ class BootstrapTable extends Component {
   }
 
   handleFilterData = filterObj => {
-    const { onFilterChange } = this.props.options;
+    const { onFilterChange, pageStartIndex } = this.props.options;
     if (onFilterChange) {
       const colInfos = this.store.getColInfos();
       onFilterChange(filterObj, colInfos);
     }
 
     this.setState({
-      currPage: this.props.options.pageStartIndex || Const.PAGE_START_INDEX,
+      currPage: Util.getFirstPage(pageStartIndex),
       reset: false
     });
 
@@ -881,7 +880,7 @@ class BootstrapTable extends Component {
 
     if (this.props.pagination) {
       const { sizePerPage } = this.state;
-      result = this.store.page(1, sizePerPage).get();
+      result = this.store.page(Util.getNormalizedPage(pageStartIndex), sizePerPage).get();
     } else {
       result = this.store.get();
     }
@@ -936,14 +935,14 @@ class BootstrapTable extends Component {
     if (this.refs.toolbar) {
       this.refs.toolbar.setSearchInput(searchText);
     }
-    const { onSearchChange } = this.props.options;
+    const { onSearchChange, pageStartIndex } = this.props.options;
     if (onSearchChange) {
       const colInfos = this.store.getColInfos();
       onSearchChange(searchText, colInfos, this.props.multiColumnSearch);
     }
 
     this.setState({
-      currPage: this.props.options.pageStartIndex || Const.PAGE_START_INDEX,
+      currPage: Util.getFirstPage(pageStartIndex),
       reset: false
     });
 
@@ -966,7 +965,7 @@ class BootstrapTable extends Component {
     let result;
     if (this.props.pagination) {
       const { sizePerPage } = this.state;
-      result = this.store.page(1, sizePerPage).get();
+      result = this.store.page(Util.getNormalizedPage(pageStartIndex), sizePerPage).get();
     } else {
       result = this.store.get();
     }
@@ -1027,7 +1026,7 @@ class BootstrapTable extends Component {
   }
 
   renderToolBar() {
-    const { exportCSV, selectRow, insertRow, deleteRow, search, children } = this.props;
+    const { exportCSV, selectRow, insertRow, deleteRow, search, children, keyField } = this.props;
     const enableShowOnlySelected = selectRow && selectRow.showOnlySelected;
     const print = typeof this.props.options.printToolBar === 'undefined' ?
       true : this.props.options.printToolBar;
@@ -1043,10 +1042,14 @@ class BootstrapTable extends Component {
       if (Array.isArray(children)) {
         columns = children.map((column, r) => {
           const { props } = column;
+          const isKey = props.isKey || keyField === props.dataField;
           return {
+            isKey,
             name: props.headerText || props.children,
             field: props.dataField,
             hiddenOnInsert: props.hiddenOnInsert,
+            keyValidator: props.keyValidator,
+            customInsertEditor: props.customInsertEditor,
             // when you want same auto generate value and not allow edit, example ID field
             autoValue: props.autoValue || false,
             // for create editor, no params for column.editable() indicate that editor for new row
@@ -1061,7 +1064,9 @@ class BootstrapTable extends Component {
           name: children.props.headerText || children.props.children,
           field: children.props.dataField,
           editable: children.props.editable,
-          hiddenOnInsert: children.props.hiddenOnInsert
+          customInsertEditor: children.props.customInsertEditor,
+          hiddenOnInsert: children.props.hiddenOnInsert,
+          keyValidator: children.props.keyValidator
         } ];
       }
       return (
@@ -1103,7 +1108,8 @@ class BootstrapTable extends Component {
             searchPanel={ this.props.options.searchPanel }
             btnGroup={ this.props.options.btnGroup }
             toolBar={ this.props.options.toolBar }
-            reset={ this.state.reset } />
+            reset={ this.state.reset }
+            isValidKey={ this.store.isValidKey } />
         </div>
       );
     } else {
@@ -1221,11 +1227,11 @@ class BootstrapTable extends Component {
       const { sizePerPage } = this.state;
 
       if (atTheBeginning) {
-        const firstPage = this.props.options.pageStartIndex || Const.PAGE_START_INDEX;
-        result = this.store.page(firstPage, sizePerPage).get();
+        const { pageStartIndex } = this.props.options;
+        result = this.store.page(Util.getNormalizedPage(pageStartIndex), sizePerPage).get();
         this.setState({
           data: result,
-          currPage: firstPage,
+          currPage: Util.getFirstPage(pageStartIndex),
           reset: false
         });
       } else {
@@ -1295,6 +1301,8 @@ BootstrapTable.propTypes = {
   insertRow: PropTypes.bool,
   deleteRow: PropTypes.bool,
   search: PropTypes.bool,
+  multiColumnSearch: PropTypes.bool,
+  strictSearch: PropTypes.bool,
   columnFilter: PropTypes.bool,
   trClassName: PropTypes.any,
   tableStyle: PropTypes.object,
@@ -1382,6 +1390,7 @@ BootstrapTable.propTypes = {
     expandRowBgColor: PropTypes.string,
     expandBy: PropTypes.string,
     expanding: PropTypes.array,
+    onlyOneExpanding: PropTypes.bool,
     beforeShowError: PropTypes.func,
     printToolBar: PropTypes.bool
   }),
@@ -1444,6 +1453,7 @@ BootstrapTable.defaultProps = {
   deleteRow: false,
   search: false,
   multiColumnSearch: false,
+  strictSearch: undefined,
   multiColumnSort: 1,
   columnFilter: false,
   trClassName: '',
@@ -1525,6 +1535,7 @@ BootstrapTable.defaultProps = {
     expandRowBgColor: undefined,
     expandBy: Const.EXPAND_BY_ROW,
     expanding: [],
+    onlyOneExpanding: false,
     beforeShowError: undefined,
     printToolBar: true
   },